// src/app/shared/types/api-types/bidt-company-holiday-dto.interface.ts
export interface BidtCompanyHolidayDTO {
  companyHolidayId?: number;
  compHolidayStartDate?: number;
  compHolidayEndDate?: number;
  compHolidayComment?: string;
  compHolidayActivated?: number;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  statusBool?: boolean;
  startDateString?: string;
  endDateString?: string;
  dataId?: number;
}